Handle empty REs at runtime
Previously empty RE were saved and reused based on their occurrence
in the compiled script.

Per POSIX if an RE is empty (that is, no pattern is specified) sed shall
behave as if the last RE used in the last command applied (either as an
address or as part of a substitute command) was specified.

Now REs are saved and reused at runtime.